Sync Fork from Upstream Repo

+++ src/link.zig
@@ -46,6 +46,7 @@ pub const Options = struct {
     entry_addr: ?u64 = null,
     stack_size_override: ?u64,
     image_base_override: ?u64,
+    include_compiler_rt: bool,
     /// Set to `true` to omit debug info.
     strip: bool,
     /// If this is true then this link code is responsible for outputting an object
@@ -57,6 +58,9 @@ pub const Options = struct {
     /// other objects.
     /// Otherwise (depending on `use_lld`) this link code directly outputs and updates the final binary.
     use_llvm: bool,
+    /// Darwin-only. If this is true, `use_llvm` is true, and `is_native_os` is true, this link code will
+    /// use system linker `ld` instead of the LLD.
+    system_linker_hack: bool,
     link_libc: bool,
     link_libcpp: bool,
     function_sections: bool,
@@ -68,6 +72,7 @@ pub const Options = struct {
     bind_global_refs_locally: bool,
     is_native_os: bool,
     pic: bool,
+    pie: bool,
     valgrind: bool,
     stack_check: bool,
     single_threaded: bool,
@@ -84,10 +89,12 @@ pub const Options = struct {
     subsystem: ?std.Target.SubSystem,
     linker_script: ?[]const u8,
     version_script: ?[]const u8,
-    override_soname: ?[]const u8,
+    soname: ?[]const u8,
     llvm_cpu_features: ?[*:0]const u8,
     /// Extra args passed directly to LLD. Ignored when not linking with LLD.
     extra_lld_args: []const []const u8,
+    /// Darwin-only. Set the root path to the system libraries and frameworks.
+    syslibroot: ?[]const u8,
 
     objects: []const []const u8,
     framework_dirs: []const []const u8,
@@ -232,7 +239,29 @@ pub const File = struct {
 
     pub fn makeExecutable(base: *File) !void {
         switch (base.tag) {
-            .coff, .elf, .macho => if (base.file) |f| {
+            .macho => if (base.file) |f| {
+                if (base.intermediary_basename != null) {
+                    // The file we have open is not the final file that we want to
+                    // make executable, so we don't have to close it.
+                    return;
+                }
+                if (comptime std.Target.current.isDarwin() and std.Target.current.cpu.arch == .aarch64) {
+                    if (base.options.target.cpu.arch != .aarch64) return; // If we're not targeting aarch64, nothing to do.
+                    // XNU starting with Big Sur running on arm64 is caching inodes of running binaries.
+                    // Any change to the binary will effectively invalidate the kernel's cache
+                    // resulting in a SIGKILL on each subsequent run. Since when doing incremental
+                    // linking we're modifying a binary in-place, this will end up with the kernel
+                    // killing it on every subsequent run. To circumvent it, we will copy the file
+                    // into a new inode, remove the original file, and rename the copy to match
+                    // the original file. This is super messy, but there doesn't seem any other
+                    // way to please the XNU.
+                    const emit = base.options.emit orelse return;
+                    try emit.directory.handle.copyFile(emit.sub_path, emit.directory.handle, emit.sub_path, .{});
+                }
+                f.close();
+                base.file = null;
+            },
+            .coff, .elf => if (base.file) |f| {
                 if (base.intermediary_basename != null) {
                     // The file we have open is not the final file that we want to
                     // make executable, so we don't have to close it.
@@ -445,52 +474,65 @@ pub const File = struct {
             break :blk full_obj_path;
         } else null;
 
+        const compiler_rt_path: ?[]const u8 = if (base.options.include_compiler_rt)
+            comp.compiler_rt_obj.?.full_object_path
+        else
+            null;
+
         // This function follows the same pattern as link.Elf.linkWithLLD so if you want some
         // insight as to what's going on here you can read that function body which is more
         // well-commented.
 
         const id_symlink_basename = "llvm-ar.id";
 
-        base.releaseLock();
+        var man: Cache.Manifest = undefined;
+        defer if (!base.options.disable_lld_caching) man.deinit();
 
-        var ch = comp.cache_parent.obtain();
-        defer ch.deinit();
+        var digest: [Cache.hex_digest_len]u8 = undefined;
 
-        try ch.addListOfFiles(base.options.objects);
-        for (comp.c_object_table.items()) |entry| {
-            _ = try ch.addFile(entry.key.status.success.object_path, null);
+        if (!base.options.disable_lld_caching) {
+            man = comp.cache_parent.obtain();
+
+            // We are about to obtain this lock, so here we give other processes a chance first.
+            base.releaseLock();
+
+            try man.addListOfFiles(base.options.objects);
+            for (comp.c_object_table.items()) |entry| {
+                _ = try man.addFile(entry.key.status.success.object_path, null);
+            }
+            try man.addOptionalFile(module_obj_path);
+            try man.addOptionalFile(compiler_rt_path);
+
+            // We don't actually care whether it's a cache hit or miss; we just need the digest and the lock.
+            _ = try man.hit();
+            digest = man.final();
+
+            var prev_digest_buf: [digest.len]u8 = undefined;
+            const prev_digest: []u8 = Cache.readSmallFile(
+                directory.handle,
+                id_symlink_basename,
+                &prev_digest_buf,
+            ) catch |err| b: {
+                log.debug("archive new_digest={} readFile error: {}", .{ digest, @errorName(err) });
+                break :b prev_digest_buf[0..0];
+            };
+            if (mem.eql(u8, prev_digest, &digest)) {
+                log.debug("archive digest={} match - skipping invocation", .{digest});
+                base.lock = man.toOwnedLock();
+                return;
+            }
+
+            // We are about to change the output file to be different, so we invalidate the build hash now.
+            directory.handle.deleteFile(id_symlink_basename) catch |err| switch (err) {
+                error.FileNotFound => {},
+                else => |e| return e,
+            };
         }
-        try ch.addOptionalFile(module_obj_path);
-
-        // We don't actually care whether it's a cache hit or miss; we just need the digest and the lock.
-        _ = try ch.hit();
-        const digest = ch.final();
-
-        var prev_digest_buf: [digest.len]u8 = undefined;
-        const prev_digest: []u8 = Cache.readSmallFile(
-            directory.handle,
-            id_symlink_basename,
-            &prev_digest_buf,
-        ) catch |err| b: {
-            log.debug("archive new_digest={} readFile error: {}", .{ digest, @errorName(err) });
-            break :b prev_digest_buf[0..0];
-        };
-        if (mem.eql(u8, prev_digest, &digest)) {
-            log.debug("archive digest={} match - skipping invocation", .{digest});
-            base.lock = ch.toOwnedLock();
-            return;
-        }
-
-        // We are about to change the output file to be different, so we invalidate the build hash now.
-        directory.handle.deleteFile(id_symlink_basename) catch |err| switch (err) {
-            error.FileNotFound => {},
-            else => |e| return e,
-        };
 
         var object_files = std.ArrayList([*:0]const u8).init(base.allocator);
         defer object_files.deinit();
 
-        try object_files.ensureCapacity(base.options.objects.len + comp.c_object_table.items().len + 1);
+        try object_files.ensureCapacity(base.options.objects.len + comp.c_object_table.items().len + 2);
         for (base.options.objects) |obj_path| {
             object_files.appendAssumeCapacity(try arena.dupeZ(u8, obj_path));
         }
@@ -498,6 +540,9 @@ pub const File = struct {
             object_files.appendAssumeCapacity(try arena.dupeZ(u8, entry.key.status.success.object_path));
         }
         if (module_obj_path) |p| {
+            object_files.appendAssumeCapacity(try arena.dupeZ(u8, p));
+        }
+        if (compiler_rt_path) |p| {
             object_files.appendAssumeCapacity(try arena.dupeZ(u8, p));
         }
 
@@ -517,15 +562,17 @@ pub const File = struct {
         const bad = llvm.WriteArchive(full_out_path_z, object_files.items.ptr, object_files.items.len, os_type);
         if (bad) return error.UnableToWriteArchive;
 
-        Cache.writeSmallFile(directory.handle, id_symlink_basename, &digest) catch |err| {
-            std.log.warn("failed to save archive hash digest file: {}", .{@errorName(err)});
-        };
+        if (!base.options.disable_lld_caching) {
+            Cache.writeSmallFile(directory.handle, id_symlink_basename, &digest) catch |err| {
+                std.log.warn("failed to save archive hash digest file: {}", .{@errorName(err)});
+            };
 
-        ch.writeManifest() catch |err| {
-            std.log.warn("failed to write cache manifest when archiving: {}", .{@errorName(err)});
-        };
+            man.writeManifest() catch |err| {
+                std.log.warn("failed to write cache manifest when archiving: {}", .{@errorName(err)});
+            };
 
-        base.lock = ch.toOwnedLock();
+            base.lock = man.toOwnedLock();
+        }
     }
 
     pub const Tag = enum {
